% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sys_file}
\alias{.sys_file}
\title{\code{system.file()} wrapper}
\usage{
.sys_file(file_path)
}
\arguments{
\item{file_path}{Character of target file relative to \verb{/inst}}
}
\description{
Use to access files in \verb{inst/}
}
