% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching.R
\name{fetch_table}
\alias{fetch_table}
\title{Query Table}
\usage{
fetch_table(conn, table, where = "", dt = TRUE, verbose = TRUE)
}
\arguments{
\item{conn}{connection to a database.}

\item{table}{a character vector naming the table to query (e.g., "arrests")}

\item{where}{a character vector for valid SQL WHERE query (e.g., "State = 'California' AND Variable = '2007'")}

\item{dt}{a logical, if true table will be coerced to \code{data.table}}

\item{verbose}{a logical, if true the function prints the SQL query}
}
\description{
Query Table
}
\examples{
\dontrun{fetch_arrests(conn = con, table = "arrests", where = "State = 'California' AND Variable = '2007'")}

}
