% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_api}
\alias{launch_api}
\title{Launch Plumber API}
\usage{
launch_api(host = "127.0.0.1", port = 591)
}
\arguments{
\item{host}{a string that is valid IPv3 or IPv6 address that is owner by this server}

\item{port}{a number that indicates the server port that should be listened on}
}
\description{
Launch Plumber API
}
